Replacement Touch-Enabled Keyboard API multitouch
Allows for access to multiple fingers at once

- the 0.9 api would only let you know about general movement of a finger.

+ This api lets you try to track a number of distinct fingers

Each event fires with an array of touches...

Each touch will be in one of the following four states:
* TouchPressed - reported for a newly recognized press -- roughly start
* TouchMoved - reported for when a press location has shifted
* TouchReleased - reported for when a press location is no longer
* TouchStationary (reported for each finger which didn't move during the
   interval that some other finger(s) triggered one of the other states

Each touch will also have an identifier. That identifier is only valid
from start through release, after release, it is free to be recycled
(although there may be multiple releases for an id) -- because of this,
you must process all releases, otherwise you may think that there was a
jump when in fact it's a different finger.

You must process all touches in the array and you must process each event,
otherwise your application will almost certainly get confused.