JI:751018 CI:
Description
Put a catch in the add to history method to ensure that the name is never null.
This has been done to match what the QT framework is pushing into the database.
We should apply this type of fix at the database level to ensure any nav service
writing to this database will be in sync so we can avoid duplicate entries.

Testing:
Deployed to target and tested: pass