JI:507002 CI:slegault
Description
In the HTML5 MediaPlayer when the user switched tuners the app would not fetch the set frequency and instead display the first frequency in the tuners range.
We now fetch the appropriate frequency when the tuner is changed.
There was also a bug in the extension where it was returning inaccurate data on updates which was providing the wrong frequency.

Testing:
Deployed to target:Pass